requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
discord.py==1.4.1
embed==0.1.0
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1
youtube-dl==2020.9.6
